+++ util/pages/user/index.js
@@ -109,6 +109,21 @@ export const getCommitsPerRepo = (repositories, slice) =>
 
 /**
  * @param {Object[]} repositories
+ * @return {Object} where every key is the `repository.name` and its value is `repositories.stargazerCount`
+ */
+export const getStarsPerRepo = (repositories, slice) =>
+  repositories
+    .concat()
+    .slice(0, slice)
+    .reduce((acc, current) => {
+      const lang = current?.node?.name;
+      acc[lang] = acc[lang] || 0;
+      acc[lang] += current.node.stargazerCount;
+      return acc;
+    }, {});
+
+/**
+ * @param {Object[]} repositories
  * @return {Object} where every key is the `primaryLanguage.name` and its value is the `primaryLanguage.color``
  */
 export const getLanguageColors = repositories =>